refactor(backup): extract Manifest contract into backup/manifest.rs
